refactor(params): narrow log-height types to u8 and encapsulate PcsParams
Log-height values (log_trace_height, log_lde_height, log_blowup,
log_folding_arity, log_final_degree) are changed from usize to u8
across all crates. These represent small logarithms (typical values
1–30) and encoding them as u8 prevents confusion between domain
sizes and their logarithms at the type level.

PcsParams fields are made pub(crate) with a validating constructor
PcsParams::new() that rejects:
- invalid folding arity (must be 1, 2, or 3)
- zero blowup (FRI requires actual blowup)
- log_final_degree + log_blowup > 64 (domain cannot exceed 2^64)
- zero queries

LiftedCoset::new()/unlifted() assert that the resulting LDE height
does not exceed 2^64, with additions widened to u16 before comparison
so the assert itself cannot overflow. FriParams::num_rounds() guards
its internal u8 addition with a debug_assert.

A shared MAX_LOG_DOMAIN_SIZE constant (= 64) is defined in
p3-miden-lifted-fri and re-used by p3-miden-lifted-stark.

Adds log2_strict_u8 helper (thin wrapper: log2_strict_usize(n) as u8)
in p3-miden-lmcs and p3-miden-lifted-air to replace log2_strict_usize
calls at u8 boundaries.

+++ p3-miden-dev-utils/src/fixtures.rs
@@ -22,7 +22,7 @@ pub const BENCH_SEED: u64 = TEST_SEED;
 // =============================================================================
 
 /// Standard log heights for benchmarking: 2^16, 2^18, 2^20 leaves.
-pub const LOG_HEIGHTS: &[usize] = &[16, 18, 20];
+pub const LOG_HEIGHTS: &[u8] = &[16, 18, 20];
 
 /// Standard relative specs for benchmark matrix groups.
 ///
